requirements.txt
flask==2.3.3
werkzeug==2.3.7
pymongo==4.1.1
python-dotenv==0.19.0
flask-cors==3.0.10 